requirements.txt
absl-py==2.1.0
dm_env==1.6
fuzzywuzzy==0.18.0
google-generativeai==0.5.1
grpcio-tools
immutabledict==2.0.0
IPython
jsonschema==4.17.3
matplotlib==3.6.1
numpy==1.26.3
opencv-python
pandas==2.1.4
pydub
python-Levenshtein
pytest
requests
tenacity
termcolor

openai==1.70.0
Pillow==11.1.0
python-dotenv==1.1.0
replicate==1.0.4
langchain_openai
langgraph
langsmith